refactor: split Python scene tools
