+++ Python/server/__init__.py
@@ -41,7 +41,13 @@ def bootstrap():
     from server import ai_navigation_tools       # noqa: F401
     from server import physics_tools              # noqa: F401
     from server import validation_tools            # noqa: F401
-    from server import scene_tools                # noqa: F401
+    from server import scene_crud_tools          # noqa: F401
+    from server import scene_sync_tools          # noqa: F401
+    from server import scene_layout_tools        # noqa: F401
+    from server import scene_procedural_tools     # noqa: F401
+    from server import scene_job_tools            # noqa: F401
+    from server import scene_nav_ai_tools         # noqa: F401
+    from server import scene_validate_tools       # noqa: F401
     from server import vroid_tools                # noqa: F401
     from server import cesium_tools               # noqa: F401
     from server import packaging_tools             # noqa: F401
